requirements.txt
GitPython==3.1.41
matplotlib==3.8.2
numpy==1.24.4
pandas==2.2.3
Pillow==10.0.1
pyarrow==13.0.0
python_box==7.2.0
scikit_learn==1.3.1
scipy==1.14.1
seaborn==0.13.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
umap_learn==0.5.6
shap==0.41.0
ipykernel